% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_codons.R
\name{mk_codons}
\alias{mk_codons}
\title{DNA codon maker/separator}
\usage{
mk_codons(dna, s = 1)
}
\value{
DNA sequence, start position
}
\description{
Takes a string of DNA bases and separates into codons/triplets
}
\examples{
mk_codons("AGTCGATAGTG", s=1)
}
