% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_dna.R
\name{random_dna}
\alias{random_dna}
\title{Pooling random DNA sequence}
\usage{
random_dna(l)
}
\value{
A string of random DNA sequence with length \code{"l"}
}
\description{
Creates a DNA sequence of \code{"l"} number of random bases
}
\examples{
random_dna(l=5)
}
